zadanie-shortener/api/requirements.txt
asgiref==3.7.2
coverage==7.4.3
Django==5.0.3
django-redis==5.4.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
factory-boy==3.3.0
hiredis==2.3.2
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
pytest==8.1.1
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-mock==3.12.0
pytz==2024.1
redis==5.0.3
sqlparse==0.4.4
typing_extensions==4.10.0
